// src/interfaces/Maquinas.ts
export interface IMaquinaWeb {
  id: number;
  product: string;
  route: string;
  id_family: number;
  id_subfamily: number;
  id_brand: number;
  model: string;
  description: string;
  info: string;
  compatibilities: string;
  image: string;
  multimedia_image: string;
  cover_page: string;
  multimedia: string;
  price: number;
  offer: number;
  offer_price: number;
  weight: number;
  min: number;
  code: string;
  visits: number;
  sales: number;
  new: number;
  delivery: number;
  registered_product: Date;
  headline: string;
  active: number;
  stock: number;
  location: string;
  packages: number;
  gift: number;
  coming_soon: number;
  pre_order: number;
  estimated_date: string;
  details: string;
  price_range: string;
  captured_by: number;
}